+++ src/Services/PropertyService/PropertyService.Application/Dtos/Properties/PropertyDetailsDto.cs
@@ -13,13 +13,13 @@ public class PropertyDetailsDto
     [JsonConverter(typeof(JsonStringEnumConverter))]
     public ListingType ListingType { get; set; }
 
-    public decimal Price { get; set; }
+    public double Price { get; set; }
 
     [JsonConverter(typeof(JsonStringEnumConverter))]
     public CurrencyUnit CurrencyUnit { get; set; }
 
     public required string Type { get; set; }
-    public decimal Area { get; set; }
+    public double Area { get; set; }
     public List<string> Images { get; set; } = [];
 
     [JsonIgnore]
