+++ src/Services/PropertyService/PropertyService.Domain/Entities/Project.cs
@@ -8,7 +8,7 @@ public class Project : AuditableEntity
     public required string Name { get; set; }
     public required string Description { get; set; }
     public Guid DeveloperId { get; set; }
-    public decimal TotalArea { get; set; }
+    public double TotalArea { get; set; }
     public Guid TotalUnits { get; set; }
     public DateTime CompletionDate { get; set; }
     public ProjectStatus Status { get; set; }
